+++ platform/src/sdk/webhooks.ts
@@ -41,7 +41,7 @@ export class Webhooks {
         );
         const url: string = baseURL.replace(/\/$/, "") + "/rules";
 
-        let [reqBodyHeaders, reqBody]: [object, any] = [{}, {}];
+        let [reqBodyHeaders, reqBody]: [object, any] = [{}, null];
 
         try {
             [reqBodyHeaders, reqBody] = utils.serializeRequestBody(req, "request", "json");